refactor props
