+++ src/components/home/sections/HomeFeature.tsx
@@ -1,5 +1,14 @@
-import { HomeFeatureProps } from "@/types/home-feature";
 import Image from "next/image";
+
+export interface HomeFeatureProps {
+  title: string;
+  description: string;
+  imagePosition: "left" | "right";
+  image: {
+    src: string;
+    alt: string;
+  };
+}
 
 const HomeFeature = ({
   title,
