--- src/types/home-feature.ts
@@ -1,9 +1,0 @@
-export interface HomeFeatureProps {
-    title: string;
-    description: string;
-    imagePosition: 'left' | 'right';
-    image: {
-        src: string;
-        alt: string;
-    };
-}
